pyproject.toml
[tool.poetry]
name = "applyr"
version = "0.1.0"
description = "Intelligent Job Market Analysis Toolkit - Scrape job postings, generate market insights, and build targeted cover letters"
authors = ["Your Name"]
readme = "README.md"
packages = [{include = "applyr"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.12.0"}
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.0"
markdown = "^3.5.0"
weasyprint = "^66.0"
pandas = "^2.0.0"
pypandoc = "^1.13"
python-docx = "^1.1.0"
cssutils = "^2.11.1"
tinycss2 = "^1.3.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
ruff = "^0.8.0"
mypy = "^1.11.0"
bandit = "^1.7.5"
pre-commit = "^3.6.0"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
responses = "^0.25.8"
freezegun = "^1.5.5"
types-pyyaml = "^6.0.12.20250915"
types-markdown = "^3.9.0.20250906"
types-requests = "^2.32.4.20250913"
types-beautifulsoup4 = "^4.12.0.20250516"

[tool.poetry.scripts]
applyr = "applyr.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=applyr --cov-report=html --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["applyr"]
omit = ["tests/*", "*/__pycache__/*", "applyr/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass"
]
precision = 2

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PIE",    # flake8-pie
    "N",      # pep8-naming
    "PL",     # pylint
    "PLR",    # pylint refactor
    "PLC",    # pylint convention
    "PLE",    # pylint error
]
ignore = [
    "E501",     # Line too long (handled by formatter)
    "B008",     # Function calls in argument defaults
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic value
    "ARG001",   # Unused function argument (for callbacks)
    "PLR0912",  # Too many branches
    "PLR0915",  # Too many statements
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["applyr"]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "S101"]  # Allow magic values and assertions in tests

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = [
    "weasyprint.*",
    "pypandoc.*",
    "docx.*",
    "cssutils.*",
    "tinycss2.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv", "__pycache__", "htmlcov", "node_modules"]
skips = ["B101", "B601", "B602"]  # Skip assert_used, shell_injection_subprocess
